Java bindings update.
- a new java example shows general jvlc use
- more functions added in InputIntf (getInputFPS now works)
- a little potential problem fixed in Playlist

+++ bindings/java/org/videolan/jvlc/InputIntf.java
@@ -42,7 +42,7 @@ public interface InputIntf {
      * currently playing playlist item.
      */
     long getInputTime();
-    long getInputPosition();
+    float getInputPosition();
     void setInputTime();
     double getInputFPS();
